requirements.txt
Flask==1.1
psycopg2-binary==2.8.6
requests==2.24.0
flake8==3.8.4
autopep8==1.5.4
nox==2020.8.22